split tensor operators, add reduction operations
