+++ gpu_tensor/lib/gpu_tensor.dart
@@ -6,5 +6,6 @@ library;
 
 export 'src/gpu_tensor_base.dart';
 export 'src/gpu_activation.dart';
-export 'src/gpu_fft.dart';
+export 'src/gpu_transform.dart';
 export 'src/gpu_data.dart';
+export 'src/gpu_operator.dart';
